_tools/requirements.txt
PyPDF2>=3.0.0
python-dotenv==1.0.0
notion-client==2.2.1
markdown==3.5.2